Only ignore syscalls if their return value is -1 (or rather < 0) AND if it is an actual event. Reason: the return value is unspecified for desired events and desired events should be notified no matter what the value of the unspecified return value is.
Still TODO: For desired events, the return value (which is unspecified)
might take any value. This might influence the PIP's state and thus
the decision. Hence, the PDP's decision might be wrong on desired events
because the signaled return value (e.g. a file descriptor) is wrong.
It needs to be figured out for which syscalls this is the case (e.g. it
probably is for open()). PIP Handler event semantics must be adjusted
accordingly.